# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.AuditAction:
    enum:
    - create
    - update
    - delete
    - login
    - logout
    - permission_grant
    - permission_revoke
    - role_assign
    - role_remove
    - export
    - import
    - api_call
    type: string
    x-enum-varnames:
    - AuditActionCreate
    - AuditActionUpdate
    - AuditActionDelete
    - AuditActionLogin
    - AuditActionLogout
    - AuditActionPermissionGrant
    - AuditActionPermissionRevoke
    - AuditActionRoleAssign
    - AuditActionRoleRemove
    - AuditActionExport
    - AuditActionImport
    - AuditActionAPICall
  domain.AuditLog:
    properties:
      action:
        $ref: '#/definitions/domain.AuditAction'
      changes:
        type: string
      companyID:
        $ref: '#/definitions/domain.CompanyID'
      createdAt:
        type: string
      entityID:
        type: string
      entityName:
        type: string
      entityType:
        type: string
      id:
        type: string
      ipaddress:
        type: string
      metadata:
        type: string
      newValues:
        type: string
      oldValues:
        type: string
      performedAt:
        type: string
      requestID:
        type: string
      userAgent:
        type: string
      userEmail:
        type: string
      userID:
        type: string
      userName:
        type: string
    type: object
...
  domain.CompanyID:
    enum:
    - all
    - gruppen
    - stalbygg
    - hybridbygg
    - industri
    - tak
    - montasje
    type: string
    x-enum-varnames:
    - CompanyAll
    - CompanyGruppen
    - CompanyStalbygg
    - CompanyHybridbygg
    - CompanyIndustri
    - CompanyTak
    - CompanyMontasje
...
  handler.AuditLogDTO:
    properties:
      action:
        type: string
      changes:
        additionalProperties: true
        type: object
      companyId:
        type: string
      entityId:
        type: string
      entityName:
        type: string
      entityType:
        type: string
      id:
        type: string
      ipAddress:
        type: string
      metadata:
        additionalProperties: true
        type: object
      newValues:
        additionalProperties: true
        type: object
      oldValues:
        additionalProperties: true
        type: object
      performedAt:
        type: string
      requestId:
        type: string
      userAgent:
        type: string
      userEmail:
        type: string
      userId:
        type: string
      userName:
        type: string
    type: object
  handler.AuditLogListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/handler.AuditLogDTO'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  handler.AuditStatsResponse:
    properties:
      actionCounts:
        additionalProperties:
          format: int64
          type: integer
        type: object
      endTime:
        type: string
      startTime:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@straye.io
    name: API Support
  description: CRM Relation API for customer, project, and sales pipeline management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Straye Relation API
  version: "1.0"
paths:
...
  /audit:
    get:
      description: Returns a paginated list of audit log entries with optional filters
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: pageSize
        type: integer
      - description: Filter by user ID
        in: query
        name: userId
        type: string
      - description: Filter by action type
        in: query
        name: action
        type: string
      - description: Filter by entity type
        in: query
        name: entityType
        type: string
      - description: Filter by entity ID
        in: query
        name: entityId
        type: string
      - description: Filter by start time (RFC3339)
        in: query
        name: startTime
        type: string
      - description: Filter by end time (RFC3339)
        in: query
        name: endTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuditLogListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List audit logs
      tags:
      - Audit
  /audit/{id}:
    get:
      description: Returns a specific audit log entry
      parameters:
      - description: Audit log ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuditLogDTO'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get audit log by ID
      tags:
      - Audit
  /audit/entity/{entityType}/{entityId}:
    get:
      description: Returns audit logs for a specific entity
      parameters:
      - description: Entity type (e.g., Customer, Project)
        in: path
        name: entityType
        required: true
        type: string
      - description: Entity ID
        in: path
        name: entityId
        required: true
        type: string
      - description: 'Maximum number of entries (default: 50)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.AuditLogDTO'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get audit logs for an entity
      tags:
      - Audit
  /audit/export:
    get:
      description: Exports audit logs for a time range as JSON
      parameters:
      - description: Start time (RFC3339)
        in: query
        name: startTime
        required: true
        type: string
      - description: End time (RFC3339)
        in: query
        name: endTime
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.AuditLog'
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Export audit logs
      tags:
      - Audit
  /audit/stats:
    get:
      description: Returns statistics about audit log actions for a time range
      parameters:
      - description: Start time (RFC3339)
        in: query
        name: startTime
        required: true
        type: string
      - description: End time (RFC3339)
        in: query
        name: endTime
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuditStatsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get audit log statistics
      tags:
      - Audit
...
securityDefinitions:
  ApiKeyAuth:
    description: API Key for system operations
    in: header
    name: x-api-key
    type: apiKey
  BearerAuth:
    description: JWT Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"